SAA-1122: Improvements
